requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
PyPDF2==1.26.0
pyttsx3==2.90